Shiny Apps
This app shows what happens to multiple regression results when there
is
considerable covariance between two contiunous predictor variables.
Although
the overall model fit does not change as the covariance is
increased (as
visualized by the regression of y onto yhat and the R
squared in the model
summary), the parameter estimates become unstable
and the confidence intervals
expand, which yields large p values even
though the relationship between the
predictors and the response variable
does not change.